+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added `Redis::monitor()` method.
 * Added `UploadedFile::getReportedFilename()` method.
 * Added `UploadedFile::getReportedMimeType()` method.
+* Added `max_filename_length` validation rule.
 
 #### Deprecations